requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
bcrypt==4.1.2